Improved home,tracing-details and tracing-settings loading behavior, less flickering. Fixed LiveData value assignment on background-thread.
