+++ Corona-Warn-App/src/deviceForTesters/java/de/rki/coronawarnapp/test/risklevel/ui/TestRiskLevelCalculationFragmentCWAViewModel.kt
@@ -3,6 +3,7 @@ package de.rki.coronawarnapp.test.risklevel.ui
 import android.content.Context
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
+import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import com.google.android.gms.nearby.exposurenotification.ExposureInformation
 import com.squareup.inject.assisted.Assisted
@@ -22,7 +23,7 @@ import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.storage.RiskLevelRepository
 import de.rki.coronawarnapp.transaction.RetrieveDiagnosisKeysTransaction
 import de.rki.coronawarnapp.transaction.RiskLevelTransaction
-import de.rki.coronawarnapp.ui.tracing.card.TracingCardViewModel
+import de.rki.coronawarnapp.ui.tracing.card.TracingCardStateProvider
 import de.rki.coronawarnapp.util.KeyFileHelper
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
 import de.rki.coronawarnapp.util.di.AppContext
@@ -31,6 +32,7 @@ import de.rki.coronawarnapp.util.ui.SingleLiveEvent
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModelFactory
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.sample
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import timber.log.Timber
@@ -45,13 +47,12 @@ class TestRiskLevelCalculationFragmentCWAViewModel @AssistedInject constructor(
     @Assisted private val handle: SavedStateHandle,
     @Assisted private val exampleArg: String?,
     @AppContext private val context: Context, // App context
-    private val dispatcherProvider: DispatcherProvider,
+    dispatcherProvider: DispatcherProvider,
     private val enfClient: ENFClient,
     private val keyCacheRepository: KeyCacheRepository,
-    private val tracingCardViewModel: TracingCardViewModel
+    tracingCardStateProvider: TracingCardStateProvider
 ) : CWAViewModel(
-    dispatcherProvider = dispatcherProvider,
-    childViewModels = listOf(tracingCardViewModel)
+    dispatcherProvider = dispatcherProvider
 ) {
 
     val startLocalQRCodeScanEvent = SingleLiveEvent<Unit>()
@@ -59,9 +60,9 @@ class TestRiskLevelCalculationFragmentCWAViewModel @AssistedInject constructor(
     val apiKeysProvidedEvent = SingleLiveEvent<DiagnosisKeyProvidedEvent>()
     val riskScoreState = MutableLiveData<RiskScoreState>(RiskScoreState())
 
-    val tracingCardState by lazy {
-        tracingCardViewModel.state
-    }
+    val tracingCardState = tracingCardStateProvider.state
+        .sample(150L)
+        .asLiveData(dispatcherProvider.Default)
 
     init {
         Timber.d("CWAViewModel: %s", this)
