+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/main/home/SubmissionCardsStateProvider.kt
@@ -1,23 +1,20 @@
 package de.rki.coronawarnapp.ui.main.home
 
+import dagger.Reusable
 import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.storage.SubmissionRepository
 import de.rki.coronawarnapp.ui.submission.ApiRequestState
 import de.rki.coronawarnapp.util.DeviceUIState
-import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
-import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.onCompletion
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.sample
 import timber.log.Timber
 import javax.inject.Inject
 
-class SubmissionCardsViewModel @Inject constructor(
-    dispatcherProvider: DispatcherProvider,
-) : CWAViewModel(dispatcherProvider = dispatcherProvider) {
+@Reusable
+class SubmissionCardsStateProvider @Inject constructor() {
 
     val state: Flow<SubmissionCardState> = combine(
         SubmissionRepository.deviceUIStateFlow,
@@ -30,8 +27,6 @@ class SubmissionCardsViewModel @Inject constructor(
         )
     }
         .onStart { Timber.v("SubmissionCardState FLOW start") }
-        .onEach { Timber.d("SubmissionCardState FLOW PRE-SAMPLE emission: %s", it) }
+        .onEach { Timber.d("SubmissionCardState FLOW emission: %s", it) }
         .onCompletion { Timber.v("SubmissionCardState FLOW completed.") }
-        .sample(150L)
-        .onEach { Timber.d("SubmissionCardState FLOW POST-SAMPLE emission: %s", it) }
 }
