+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/util/ui/ViewExtensions.kt
@@ -8,6 +8,7 @@ fun View.setGone(gone: Boolean) {
     visibility = if (gone) View.GONE else View.VISIBLE
 }
 
+@BindingAdapter("invisible")
 fun View.setInvisible(invisible: Boolean) {
     visibility = if (invisible) View.INVISIBLE else View.VISIBLE
 }
